make call to EmbedBuilder#builder not required
