+++ src/commands/config/mod-logs.ts
@@ -21,7 +21,7 @@ export async function onSlashCommand(client: FluorineClient, interaction: ChatIn
             value
         });
 
-    interaction.reply({ embeds: [embed.builder] });
+    interaction.reply({ embeds: [embed] });
 }
 
 export const slashCommandData = new SlashCommandSubcommandBuilder('MOD_LOGS').addBooleanOption('MOD_LOGS', option =>
