+++ src/commands/deploy/create.ts
@@ -65,7 +65,7 @@ export async function onSlashCommand(client: FluorineClient, interaction: ChatIn
             .setTitle('Failed', { raw: true })
             .setDescription(`\`\`\`js\n${error}\`\`\``, { raw: true });
 
-        interaction.editReply({ embeds: [embed.builder] });
+        interaction.editReply({ embeds: [embed] });
     }
 }
