+++ src/commands/dev/eval.ts
@@ -53,7 +53,7 @@ export async function onModal(
         embed.setTitle('Failed', { raw: true }).setDescription(codeBlock('js', cleaned), { raw: true });
     }
 
-    interaction.reply({ embeds: [embed.builder] });
+    interaction.reply({ embeds: [embed] });
 }
 
 export const modalData: NonCommandInteractionData = {
