+++ src/commands/info.ts
@@ -104,7 +104,7 @@ export async function onInteraction(
     );
 
     const options = {
-        embeds: [embed.builder],
+        embeds: [embed],
         components: [row]
     };
