+++ src/commands/serverinfo.ts
@@ -27,7 +27,7 @@ export async function onSlashCommand(client: FluorineClient, interaction: ChatIn
         }
     );
 
-    interaction.reply({ embeds: [embed.builder] });
+    interaction.reply({ embeds: [embed] });
 }
 
 export const slashCommandData = new SlashCommandBuilder('SERVERINFO').setDMPermission(false);
