+++ src/commands/slut.ts
@@ -12,7 +12,7 @@ export async function onSlashCommand(client: FluorineClient, interaction: ChatIn
         });
 
     client.economy.add(interaction.guildId, interaction.user, money);
-    interaction.reply({ embeds: [embed.builder] });
+    interaction.reply({ embeds: [embed] });
 }
 
 export const slashCommandData = new SlashCommandBuilder('SLUT').setDMPermission(false);
