+++ src/commands/timeout.ts
@@ -71,7 +71,7 @@ export async function onSlashCommand(client: FluorineClient, interaction: ChatIn
             { name: 'CASE_ID', rawValue: caseObj.caseId.toString() }
         );
 
-    interaction.reply({ embeds: [embed.builder] });
+    interaction.reply({ embeds: [embed] });
     client.cases.logToModerationChannel(interaction.guildId, caseObj);
 }
