+++ src/types/index.ts
@@ -1,3 +1,4 @@
+export * from './builders.js';
 export * from './hypixel.js';
 export * from './structures.js';
 export * from './webRequests.js';
